lab/requirements.txt
certifi==2017.4.17
chardet==3.0.4
colorama==0.3.9
decorator==4.0.11
diff-match-patch==20121119
Django==1.11.2
django-browserid==2.0.2
django-classy-tags==0.8.0
django-import-export==0.5.1
django-missing==0.1.18
django-mongo-auth==0.1.3
django-sekizai==0.10.0
et-xmlfile==1.0.1
idna==2.5
ipython==6.1.0
ipython-genutils==0.2.0
jdcal==1.3
jedi==0.10.2
mongoengine==0.13.0
mysqlclient==1.3.10
oauthlib==2.0.2
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.9
pickleshare==0.7.4
Pillow==4.2.0
prompt-toolkit==1.0.14
Pygments==2.2.0
pymongo==3.4.0
PyMySQL==0.7.11
pytz==2017.2
PyYAML==3.12
requests==2.18.1
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.10.0
tablib==0.12.1
testools==0.0.1
traitlets==4.3.2
tweepy==3.5.0
unicodecsv==0.14.1
urllib3==1.21.1
wcwidth==0.1.7
win-unicode-console==0.5
xlrd==1.1.0
xlwt==1.3.0
